+++ test/integration/clitest/test_render.py
@@ -19,15 +19,12 @@
 # with this program; if not, write to the Free Software Foundation, Inc.,
 # 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 
-from __future__ import division
-from builtins import str
-from builtins import range
-from past.utils import old_div
 import json
 import pytest
 
 from omero_cli_render import RenderControl
 from omero.cli import NonZeroReturnCode
+from omero.plugins.delete import DeleteControl
 from cli import CLITest
 from omero.gateway import BlitzGateway
 import os.path
@@ -43,6 +40,8 @@ class TestRender(CLITest):
     def setup_method(self, method):
         super(TestRender, self).setup_method(method)
         self.cli.register("render", RenderControl, "TEST")
+        self.cli.register("delete", DeleteControl, "TEST")
+        self.delete_args = self.args + ["delete"]
         self.args += ["render"]
         self.idonly = "-1"
         self.imageid = "Image:-1"
@@ -205,7 +204,7 @@ class TestRender(CLITest):
                 assert img.getDefaultZ() == rdef.get('z') - 1
             else:
                 # If not set, default Z plane is the middle one
-                assert img.getDefaultZ() == (int)(old_div(img.getSizeZ(), 2))
+                assert img.getDefaultZ() == (int)(img.getSizeZ() // 2)
 
     def assert_channel_rdef(self, channel, rdef, version=2):
         assert channel.getLabel() == rdef['label']
@@ -416,3 +415,13 @@ class TestRender(CLITest):
         self.args += ["set", self.idonly, str(rdfile)]
         self.cli.invoke(self.args, strict=True)
         self.assert_target_rdef(self.idonly, rd)
+
+    def test_thumb_no_renderingdef(self):
+        """Test thumbnail generation when the Image has no RenderingDef"""
+        self.create_image()
+        self.delete_args += [
+            "Image/Thumbnail:" + self.idonly,
+            "Image/RenderingDef:" + self.idonly]
+        self.cli.invoke(self.delete_args, strict=True)
+        self.args += ["test", "--thumb", self.idonly]
+        self.cli.invoke(self.args, strict=True)